+++ doc/whats-new.rst
@@ -23,6 +23,8 @@ v0.17.1 (unreleased)
 New Features
 ~~~~~~~~~~~~
 
+- Allow assigning values to a subset of a dataset using positional or label-based
+  indexing (:issue:`3015`). By `Matthias Göbel <https://github.com/matzegoebel>`_.
 - Add :py:meth:`Dataset.query` and :py:meth:`DataArray.query` which enable indexing
   of datasets and data arrays by evaluating query expressions against the values of the
   data variables (:pull:`4984`). By `Alistair Miles <https://github.com/alimanfoo>`_.